backend/functions/requirements.txt
e2b==0.17.1
e2b-code-interpreter==0.0.10a0
firebase_functions~=0.1.0
gcloud
google-cloud-firestore
google-cloud-storage
firebase-admin
flask
fpdf
openAI
werkzeug
